requirements.txt
ckanapi
libhxl>=1.7
